# docs/swagger.yaml
basePath: /api/v1
definitions:
  helpers.HTTPError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  models.ResponseMessage:
    properties:
      message:
        example: runner successfully deleted
        type: string
    type: object
  models.Role:
    properties:
      access:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      resource:
        type: string
      resource_ids:
        items:
          type: string
        type: array
        uniqueItems: true
      updated_at:
        type: string
    required:
    - access
    - name
    - resource
    - resource_ids
    type: object
  models.RoleBinding:
    properties:
      created_at:
        type: string
      group_id:
        type: string
      id:
        type: string
      name:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
...
info:
  contact:
    email: info@kubecode.io
    name: Kubecode Support
    url: https://www.kubecode.io/contact
  description: API Gateway for your kubernetes services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Kriten
  version: v0.3
paths:
...
  /role_bindings:
    get:
      consumes:
      - application/json
      description: List all roles bindings available on the cluster
      parameters:
      - description: Maximum number of role bindings to return (default 100)
        in: query
        name: limit
        type: integer
      - description: Number of role bindings to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by role binding name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RoleBinding'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: List all role bindings
      tags:
      - rolebindings
    post:
      consumes:
      - application/json
      description: Add a role binding to the cluster
      parameters:
      - description: New role binding
        in: body
        name: roleBinding
        required: true
        schema:
          $ref: '#/definitions/models.RoleBinding'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoleBinding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Create a new role binding
      tags:
      - rolebindings
  /role_bindings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete by role binding ID
      parameters:
      - description: RoleBinding ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Delete a role binding
      tags:
      - rolebindings
    get:
      consumes:
      - application/json
      description: Get information about a specific role binding
      parameters:
      - description: RoleBinding ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoleBinding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Get a role binding
      tags:
      - rolebindings
    patch:
      consumes:
      - application/json
      description: Update a role binding in the cluster
      parameters:
      - description: RoleBinding ID
        in: path
        name: id
        required: true
        type: string
      - description: Update role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.RoleBinding'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoleBinding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Update a role binding
      tags:
      - rolebindings
  /roles:
    get:
      consumes:
      - application/json
      description: List all roles available on the cluster
      parameters:
      - description: Maximum number of roles to return (default 100)
        in: query
        name: limit
        type: integer
      - description: Number of roles to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by role name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: List all roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Add a role to the cluster
      parameters:
      - description: New role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Create a new role
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete by role ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Delete a role
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Get information about a specific role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Get a role
      tags:
      - roles
    patch:
      consumes:
      - application/json
      description: Update a role in the cluster
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Update role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Update a role
      tags:
      - roles
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"